requirements.txt
Django==1.8.2
django-crispy-forms==1.4.0
python-social-auth==0.2.11
